Reworked initialization of any and all - Now initialized by a single function call - Moved boolean reduction template into header
